Fix #8
It seems a work everywhere, sensible and practical solution is not
feasible. So, we enable the users to customize the name generation and
provide a reasonable default, which works for the 2 main online
judges (at least on normal contests): Codeforces and AtCoder.

However, I am open to better suggestions in the future.